Restrict topics routes to those that are needed
This makes the output of rake routes more concise, and produces better
error messages when someone accidentially visits one of those URLs.

It might make sense to go over the other resources in the routes file as
well and check what can be further restricted.

Also remove the new action from TopicsController which is not used.

Refs #295 by @acrogenesis which accidentially fell victim to a
misinterpretation on my part. Mea culpa.