Format code in markdown and haml CoC by wrapping text lines
This wraps the text in the markdown and haml code by adding newlines
every 80 chars or so. It makes the code easier to read when viewed with
plain text tools. Especially the markdown format is intended to be well
formatted and readable as plain text, so I think it makes sense there.
For haml or html code in general it is optional, and has no effect on
the way it is rendered in the browser, but still a nice touch imho.