Fix gitignore of the secrets.yml
Apparently a comment is not allowed at the end of the line. Also the
comment doesn't say a lot, because probably not only shelly needs it
that way, so I was so free to just delete the comment.

Anyway, now you can have a local config/secrets.yml without risking to
commit it to git.