Turn off geocoding in Development (#685)
* Remove Geocoding validation from development

After some testing, we found that if a user copied the secrets.yml.example
before seeding the database OR creating a group through the interface,
the geocoding validations on the group would cause errors. For this reason,
we are turning it off for development. We wanted to keep the tests running
with geocoder, so this is only a fix for development. It would be great to add it
back in one day, if we can figure out how to avoid problem in development

Co-authored-by: Franka Schmidt <hello@franka.tech>
Co-authored-by: Sara Regan <regan.sara@gmail.com>

* Add comment to explain turning off geocode validation

Co-authored-by: Franka Schmidt <hello@franka.tech>
Co-authored-by: Sara Regan <regan.sara@gmail.com>